backend/requirements-windows.txt
# Windows-Compatible Requirements
# Install these in order

# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
aiofiles==23.2.1

# LlamaIndex core
llama-index-core==0.11.20

# Ollama integrations
llama-index-llms-ollama==0.3.4
llama-index-embeddings-ollama==0.3.1

# Vector store - Using simpler alternative for Windows
llama-index-vector-stores-qdrant==0.3.2
qdrant-client==1.12.1

# OR if you have Visual Studio C++ Build Tools:
# chromadb==0.5.23
